+++ src/Renderer/RendererInterface.php
@@ -4,4 +4,6 @@ namespace FOA\Responder_Bundle\Renderer;
 interface RendererInterface
 {
     public function render($data, $view, $layout = null);
+
+    public function getEngine();
 }