+++ src/Renderer/Twig.php
@@ -16,4 +16,9 @@ class Twig implements RendererInterface
     {
         return $this->engine->render($view, $data);
     }
+
+    public function getEngine()
+    {
+        return $this->engine;
+    }
 }