+++ tests/integration/src/MustacheResponderTest.php
@@ -38,4 +38,14 @@ class MustacheResponderTest extends \PHPUnit_Framework_TestCase
         $this->responder->__invoke();
         $this->assertSame('Hello Hari', trim($this->response->content->get()));
     }
+
+    public function testGetRenderer()
+    {
+        $this->assertInstanceOf('FOA\Responder_Bundle\Renderer\Mustache', $this->responder->getRenderer());
+    }
+
+    public function testGetEngine()
+    {
+        $this->assertInstanceOf('Mustache_Engine', $this->responder->getRenderer()->getEngine());
+    }
 }