+++ tests/integration/src/PlatesResponderTest.php
@@ -35,4 +35,14 @@ class PlatesResponderTest extends \PHPUnit_Framework_TestCase
         $this->responder->__invoke();
         $this->assertSame('Hello Hari', trim($this->response->content->get()));
     }
+
+    public function testGetRenderer()
+    {
+        $this->assertInstanceOf('FOA\Responder_Bundle\Renderer\Plates', $this->responder->getRenderer());
+    }
+
+    public function testGetEngine()
+    {
+        $this->assertInstanceOf('League\Plates\Engine', $this->responder->getRenderer()->getEngine());
+    }
 }