+++ tests/integration/src/TwigResponderTest.php
@@ -37,4 +37,14 @@ class TwigResponderTest extends \PHPUnit_Framework_TestCase
         $this->responder->__invoke();
         $this->assertSame('Hello Hari', trim($this->response->content->get()));
     }
+
+    public function testGetRenderer()
+    {
+        $this->assertInstanceOf('FOA\Responder_Bundle\Renderer\Twig', $this->responder->getRenderer());
+    }
+
+    public function testGetEngine()
+    {
+        $this->assertInstanceOf('Twig_Environment', $this->responder->getRenderer()->getEngine());
+    }
 }